update marging triggered by data change instead of settings change
