+++ visualization/app/codeCharta/core/scenario/scenario.service.ts
@@ -2,7 +2,6 @@
 import {Settings, SettingsService} from "../settings/settings.service";
 import {createDefaultScenario} from "./scenario.data";
 import {ThreeOrbitControlsService} from "../../ui/codeMap/threeViewer/threeOrbitControlsService";
-import {CodeMapNode} from "../data/model/CodeMap";
 
 export interface Scenario {
     name: string;
@@ -22,13 +21,6 @@ export class ScenarioService {
                 private dataService,
                 private threeOrbitControlsService:ThreeOrbitControlsService) {
         this.scenarios = require("./scenarios.json");
-        this.settingsService.subscribe(this);
-    }
-
-
-    public onSettingsChanged(){
-        this.settingsService.settings.margin = this.computeMargin();
-        this.threeOrbitControlsService.autoFitTo();
     }
 
     /**
@@ -41,76 +33,6 @@ export class ScenarioService {
             this.threeOrbitControlsService.autoFitTo();
         }
     }
-
-    /**
-     * @returns {number}
-     *
-     * Function that computes the margin applied to a scenario related the square root of the area divided
-     * by the number of buildings
-     */
-    public  computeMargin(): number{
-
-        let root: CodeMapNode = this.dataService._data.renderMap.root;
-        let margin: number;
-
-        // Previous parameters are wrapped to introduce them into goThroughMap() and make value changes
-        // remain, with a pointer-like behaviour
-        let numberOfBuildingsWrapped = {numberOfBuildings:0};
-         let areaParametersWrapped = {areaMetricName: this.settingsService.settings.areaMetric, totalArea: 0};
-
-        this.dataService.goThroughMap( root, this.countBuildings, numberOfBuildingsWrapped);
-        this.dataService.goThroughMap( root, this.calculateTotalArea, areaParametersWrapped);
-
-        margin = 4*Math.round(
-            Math.sqrt(
-                (areaParametersWrapped["totalArea"]/numberOfBuildingsWrapped["numberOfBuildings"])));
-
-        console.log("buildings "+numberOfBuildingsWrapped["numberOfBuildings"]+" area "+
-            areaParametersWrapped["areaMetricName"] +" "+areaParametersWrapped["totalArea"]+" margin "+margin);
-        return margin;
-    }
-
-    /*
-    * @param {CodeMapNode} map
-    * @param {number[]} buildings ["numberOfBuildings"] the number of leaves( aka buildings) in map
-    *
-     */
-    public countBuildings(map: CodeMapNode, buildings: number[]){
-        if(!map.children){
-            buildings["numberOfBuildings"]++;
-        }
-    }
-
-    /**
-     *
-     * @param map
-     * @param area ["areaMetricName"] name of the metric used for area in the map,
-     *          ["totalArea"] aggregated value of the area
-     */
-    public calculateTotalArea(map: CodeMapNode, area: any[]){
-        if(!map.children){
-            area["totalArea"]+=map.attributes[area["areaMetricName"]];
-        }
-    }
-
-    /**
-     *
-     * @param {CodeMapNode} map
-     * @param {number[]} deep
-     *
-     * Function that called by goThroughMap() finds the deeper level of a map.
-     */
-    public findMapDeep(map:CodeMapNode,deep: number[]){
-        if(map.children){
-            deep["currentDeep"]++;
-        }
-        else{
-            deep["maximalDeep"]=Math.max(deep["currentDeep"]+1,deep["maximalDeep"]);
-            if(deep["currentDeep"]==342) console.log("deepName: "+map.path);
-            deep["currentDeep"]=0;
-        }
-    }
-
 
     /**
      * Returns an array of all scenarios.
@@ -134,7 +56,7 @@ export class ScenarioService {
      * @returns {Scenario} the scenario
      */
     public getDefaultScenario(): Scenario {
-        return createDefaultScenario(this.settingsService.settings.map, this.computeMargin());
+        return createDefaultScenario(this.settingsService.settings.map, this.settingsService.computeMargin());
     }
 
 }
